+++ packages/zeebe-element-templates-json-schema/test/fixtures/additional-property.js
@@ -1,0 +1,11 @@
+export const template = {
+  'name': 'AdditionalProperty',
+  'id': 'com.camunda.example.AdditionalProperty',
+  'appliesTo': [
+    'bpmn:Task'
+  ],
+  'properties': [],
+  'foo': 'bar'
+};
+
+export const errors = null;