+++ packages/zeebe-element-templates-json-schema/test/fixtures/invalid-version.js
@@ -1,0 +1,33 @@
+export const template = {
+  'name': 'InvalidVersion',
+  'id': 'com.camunda.example.InvalidVersion',
+  'version': 'foo',
+  'appliesTo': [
+    'bpmn:Task'
+  ],
+  'properties': []
+};
+
+export const errors = [
+  {
+    keyword: 'type',
+    dataPath: '/version',
+    schemaPath: '#/allOf/0/properties/version/type',
+    params: { type: 'number' },
+    message: 'should be number'
+  },
+  {
+    keyword: 'type',
+    dataPath: '',
+    schemaPath: '#/oneOf/1/type',
+    params: { type: 'array' },
+    message: 'should be array'
+  },
+  {
+    keyword: 'oneOf',
+    dataPath: '',
+    schemaPath: '#/oneOf',
+    params: { passingSchemas: null },
+    message: 'should match exactly one schema in oneOf'
+  }
+];