+++ packages/zeebe-element-templates-json-schema/test/fixtures/missing-type.js
@@ -1,0 +1,18 @@
+export const template = {
+  'name': 'MissingType',
+  'id': 'com.camunda.example.MissingType',
+  'appliesTo': [
+    'bpmn:Task'
+  ],
+  'properties': [
+    {
+      'label': 'foo',
+      'binding': {
+        'type': 'property',
+        'name': 'foo'
+      }
+    }
+  ]
+};
+
+export const errors = null;