+++ packages/zeebe-element-templates-json-schema/test/fixtures/with-version.js
@@ -1,0 +1,11 @@
+export const template = {
+  'name': 'WithVersion',
+  'id': 'com.camunda.example.WithVersion',
+  'version': 200,
+  'appliesTo': [
+    'bpmn:Task'
+  ],
+  'properties': []
+};
+
+export const errors = null;