+++ packages/zeebe-element-templates-json-schema/test/helpers/index.js
@@ -1,0 +1,13 @@
+import Ajv from 'ajv';
+
+export function createValidator(schema) {
+
+  const ajv = new Ajv({
+    allErrors: true,
+    strict: false
+  });
+
+  require('ajv-errors')(ajv);
+
+  return ajv.compile(schema);
+}